Fix some old bugs: make title texture a power of 2, other small fixes that were never commited.

+++ src/renderer.h
@@ -5,7 +5,7 @@
 #include <vector>
 
 typedef struct Vertex{
-	short vertx, verty;
+	float vertx, verty;
 	short textx, texty;
 	unsigned int texID;
 } Vertex;
